+++ bin/jimp.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { runCliProcess } from "../compiler/src/cli.js";
+
+await runCliProcess();